+++ packages/experiment-tag/src/util/messenger.ts
@@ -1,6 +1,6 @@
 import { getGlobalScope } from '@amplitude/experiment-core';
 
-import { getStorageItem } from '../storage/storage';
+import { getStorageItem } from '../storage/consent-aware-storage';
 
 interface VisualEditorSession {
   injectSrc: string;
